+++ example/string_lower.dart
@@ -1,0 +1,9 @@
+import 'package:support/string.dart';
+
+///
+/// Example of using the `lower` extension method on String.
+///
+void main() {
+  print("FOO BAR BAZ".lower()); // Output: "foo bar baz"
+  print("foO bAr BaZ".lower()); // Output: "foo bar baz"
+}